refactor: move claude compat code to leaf helpers
